# Documentação/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/tarefas/{id}":{"get":{"tags":["tarefa-controller"],"operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tarefa"}}}}}},"put":{"tags":["tarefa-controller"],"operationId":"atualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tarefa"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tarefa"}}}}}},"delete":{"tags":["tarefa-controller"],"operationId":"deletar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/quadros/{id}":{"get":{"tags":["quadro-controller"],"operationId":"buscarPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Quadro"}}}}}},"put":{"tags":["quadro-controller"],"operationId":"atualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quadro"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Quadro"}}}}}},"delete":{"tags":["quadro-controller"],"operationId":"deletar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/tarefas":{"get":{"tags":["tarefa-controller"],"operationId":"listarTodas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tarefa"}}}}}}},"post":{"tags":["tarefa-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tarefa"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tarefa"}}}}}},"patch":{"tags":["tarefa-controller"],"operationId":"mover","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoverTarefaDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/quadros":{"get":{"tags":["quadro-controller"],"operationId":"listarTodosPorUsuario","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Quadro"}}}}}}},"post":{"tags":["quadro-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quadro"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Quadro"}}}}}}},"/api/colunas":{"get":{"tags":["coluna-controller"],"operationId":"listarColunas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Coluna"}}}}}}},"post":{"tags":["coluna-controller"],"operationId":"criarColuna","parameters":[{"name":"coluna","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Coluna"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Coluna"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Coluna":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nome":{"type":"string","minLength":1}},"required":["nome"]},"Quadro":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string","minLength":1},"status":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"},"usuario":{"$ref":"#/components/schemas/Usuario"}},"required":["nome","usuario"]},"Tarefa":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descricao":{"type":"string","minLength":1},"dataCriacao":{"type":"string","format":"date-time"},"prazo":{"type":"string","format":"date-time"},"quadro":{"$ref":"#/components/schemas/Quadro"},"coluna":{"$ref":"#/components/schemas/Coluna"}},"required":["descricao"]},"Usuario":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string","format":"email","minLength":1},"username":{"type":"string","minLength":1}},"required":["email","username"]},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","password","username"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"MoverTarefaDto":{"type":"object","properties":{"idTarefa":{"type":"integer","format":"int32"},"idColuna":{"type":"integer","format":"int32"}}}}}}